Refactor and enhance R-Q-Evolve scripts and configurations
- Updated `merge_fsdp_to_hf.py` to use `torch_dtype` instead of `dtype` for model configuration.
- Modified `run_train_rq_evolve_base_4b.sh` to include an optional export for `CUDA_VISIBLE_DEVICES`.
- Added new configuration files for R-Q-Evolve with varying `num_rollouts` (5, 20, 40).
- Introduced `collect_scores.py` to aggregate benchmark scores into a markdown table.
- Created `eval_steps_fanout.sh` for evaluating checkpoints across multiple benchmarks in parallel.
- Implemented `gpt_recheck_only.py` to perform offline GPT-4o re-checks on evaluation results.
- Developed `make_sweep_config.py` to generate per-run configurations for sweep experiments.
- Added `sweep_num_rollouts.sh` to facilitate sweeping over different `num_rollouts` values while maintaining output consistency.

+++ scripts/merge_fsdp_to_hf.py
@@ -96,7 +96,7 @@ def main():
     print("[merge] building HF model skeleton...")
     config = AutoConfig.from_pretrained(hf_src)
     with torch.device("meta"):
-        model = AutoModelForCausalLM.from_config(config, dtype=target_dtype)
+        model = AutoModelForCausalLM.from_config(config, torch_dtype=target_dtype)
     model.to_empty(device="cpu")
 
     missing, unexpected = model.load_state_dict(full_sd, strict=False, assign=True)
